requirements.txt
numpy==1.26.1
opencv-python==4.8.1.78
PyYAML==6.0.1
UE4Parse @ git+https://github.com/MinshuG/pyUE4Parse.git@4714f08d011445f959f51aa13865f6641f78c5d6
